requirements.txt
flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.1.0
